+++ CHANGES.md
@@ -11,6 +11,8 @@ Unreleased changes refer to our current [master branch](https://github.com/openc
 * Automatic Outlook inbox refresh when receiving new emails
 
 ### Fixes
+* Deny the removal of a special folder
+* Reuse special folders if a new one is being created with the same name
 * No more `Deleted Items (1)`-like duplicated folders
 * Fixed creation of root folders on online mode and some special folders such as Sync Issues.
 * Fixed `Too many connections to ldap` when openchange runs on samba as member of a domain.